Re-throw AsyncException in performTestCase
Ignore-this: 9b4cfb4238bef682c4e6d30625743625

This is important, as otherwise test runs will not terminate on SIGINT
(ctrl-c).  We could re-throw the exception only it is a UserInterrupt, but the
other asynchronous exceptions indicate severe conditions, so I think it is a
good idea to re-throw all of them.

darcs-hash:20110809104735-ec922-4f803c9316f2955df4900256daa80d414bca7a21