Commented out some tests due to Hackage conflict
Ignore-this: 369730ad000c6d7d97a05031b665dd96

Some of the package tests verify that tests are not optimized out
of existence at the various optimization levels (-O0, -O1, -O2).  By
necessity, these tests must be compiled with different levels of
optimization.  Cabal complains when optimization levels are manually
specified and Hackage won't accept packages that Cabal complains about.

As a result, some of the tests are temporarily disabled by this patch.
The tests should still catch the issue on 6.12.x and 7.0.x (the currently
available versions); however, it will be desireable to reinstate these
tests through some trickery.  E.g., use custom hooks to trigger the building
of a second Cabal file, which builds and runs the tests.

This is going up as-is because it solves 2 major issues (compared to 1.2).

darcs-hash:20110813212129-918ee-3b47feb3ce098b63c854e9126314d81afe29f59a